+++ src/app/components/share-show/-header-base.js
@@ -27,7 +27,7 @@ export default Component.extend(I18n, {
   /**
    * One of: share, handle.
    * If share - it is a link to Onezone's share.
-   * If handle - it is a link to published Open Data (in handle service).
+   * If handle - it is a link to exposed Public Data (in handle service).
    * Initialized in init, if left undefined.
    * @type {String}
    */