+++ src/app/locales/en/components/share-show.js
@@ -3,11 +3,11 @@ import headerPublic from './share-show/header-public';
 import publicUrlViewer from './share-show/public-url-viewer';
 import paneDescription from './share-show/pane-description';
 import paneFiles from './share-show/pane-files';
-import paneOpendata from './share-show/pane-opendata';
+import paneOpendata from './share-show/pane-publicdata';
 import dcCommon from './share-show/dc-common';
 import dcEditor from './share-show/dc-editor';
 import dcPreview from './share-show/dc-preview';
-import openDataModeSwitch from './share-show/open-data-mode-switch';
+import publicDataModeSwitch from './share-show/public-data-mode-switch';
 import markdownEditor from './share-show/markdown-editor';
 import urlInfoHandle from './share-show/url-info-handle';
 import urlInfoShare from './share-show/url-info-share';
@@ -17,11 +17,11 @@ import metadataEditorFooter from './share-show/metadata-editor-footer';
 
 export default {
   tabs: {
-    opendata: 'Open Data',
+    publicdata: 'Public Data',
     description: 'Description',
     files: 'Files',
     filesDeleted: 'Files <span class="text-danger">(deleted)</span>',
-    publishOpenData: 'Publish as Open Data',
+    exposePublicData: 'Expose as Public Data',
   },
   // used accross misc components
   modifyButton: 'Modify',
@@ -38,7 +38,7 @@ export default {
     editor: dcEditor,
     preview: dcPreview,
   },
-  openDataModeSwitch,
+  publicDataModeSwitch,
   markdownEditor,
   urlInfoHandle,
   urlInfoShare,