+++ src/app/locales/en/components/share-show/url-info-common.js
@@ -1,3 +1,3 @@
 export default {
-  selectorInfo: 'Use the dropdown on the left to switch between different public links.',
+  selectorInfo: 'Use the dropdown on the left to switch between different links.',
 };